rpi/requirements_rpi.txt
# Raspberry Pi optimized requirements
# Full TensorFlow for Raspberry Pi (ARM64)
tensorflow>=2.10.0

# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Web framework
flask>=2.0.0
gunicorn>=21.2.0

# GPIO and sensor libraries
RPi.GPIO>=0.7.0
adafruit-circuitpython-dht>=3.7.0
adafruit-circuitpython-ads1x15>=3.3.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Additional utilities
psutil>=5.8.0
requests>=2.25.0
python-dotenv>=0.19.0

# For real-time data logging
sqlite3
csv 